requirements.txt
ccxt>=4.0
pandas>=2.0
pandas_ta>=0.3.14b
numpy>=1.24
xgboost>=2.0
scikit-learn>=1.3
joblib>=1.3
openai>=1.0
python-dotenv>=1.0
rich>=13.0
